feat: shell script -> commander
